+++ src/clj_kondo/core.clj
@@ -21,8 +21,10 @@
 (defn print!
   "Prints the result from `run!` to `*out*`. Returns `nil`. Alpha,
   subject to change."
-  [{:keys [:config :findings :summary :analysis]}]
+  [{:keys [:config :findings :summary :analysis]} report-level]
   (let [output-cfg (:output config)
+        report-level? (set (drop-while #(not= (keyword report-level) %)
+                                       [:info :warning :error]))
         fmt (or (:format output-cfg) :text)]
     (case fmt
       :text
@@ -31,11 +33,15 @@
                                        (println)))
         (let [format-fn (core-impl/format-output config)]
           (doseq [finding findings]
-            (println (format-fn finding)))
+            (when (report-level? (:level finding))
+              (println (format-fn finding))))
           (when (:summary output-cfg)
             (let [{:keys [:error :warning :duration]} summary]
               (printf "linting took %sms, " duration)
-              (println (format "errors: %s, warnings: %s" error warning))))))
+              (printf "errors: %s" error)
+              (when (report-level? "warning")
+                (printf ", warnings: %s" warning))
+              (println "")))))
       ;; avoid loading clojure.pprint or bringing in additional libs for printing to EDN for now
       :edn
       (let [output (cond-> {:findings findings}